refactor(mobile): keep the expo-router tree to routes only
Move the eight remaining non-route modules from mobile/app into their existing feature directories under mobile/src. Update consumers and add a route-tree guard so Expo Router only sees real routes.
