+++ mobile/src/session/mobile-session-command-input-styles.ts
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native'
 
-import { colors, spacing, radii, typography } from '../../../../src/theme/mobile-theme'
+import { colors, spacing, radii, typography } from '../theme/mobile-theme'
 
 export const mobileSessionCommandInputStyles = StyleSheet.create({
   createWarningBanner: {
