+++ mobile/src/terminal/terminal-live-input-affordance.test.ts
@@ -10,7 +10,7 @@ const liveInputStatusSource = readFileSync(
   'utf8'
 )
 const commandInputStylesSource = readFileSync(
-  new URL('../../app/h/[hostId]/session/mobile-session-command-input-styles.ts', import.meta.url),
+  new URL('../session/mobile-session-command-input-styles.ts', import.meta.url),
   'utf8'
 )
